minor refactoring of serialization
